FrontEnd Commit 1
Hi Craig and Team!

I have created a shelled front-end to this build that includes the following aside from the default ASP.NET packages.

App_Data currrently contains a dummy database to store the values we are required to for each app.

Default.aspx has been modified to display any apps present in the database. It will need input validiation validation eventually such that an app store will no results will display "No results"

There is a button on the top toolbar that takes the user to the App Submit page from Site.Master.

SubmitApp.aspx is a page that I wrote that has all the text boxes required for generating a new app request. This will need input validiation, and the button on this page is linked to a piece of dummy code in SubmitApp.aspx.cs
This method, named "SubmitButton_Click" will eventually grab the values stored in the textboxes and store them in a SQL database.

Craig, if you can tackle handling the SubmitButton_Click method, and create the scripts for the remainder of the Tables we are going to need this project, I think it will be sufficient with this iteration. Check with Keegan.